Provide "gcc" as argument to Boost's bootstrap.bat when using MinGW
There is a long standing bug in Boost where people may experience the
following when running bootstrap.bat:

Building Boost.Build engine
******  B A T C H   R E C U R S I O N  exceeds STACK limits ******
Recursion Count=593, Stack Usage=90 percent
******       B A T C H   PROCESSING IS   A B O R T E D      ******

This occurs when bootstrap.bat is unable to find a compiler toolchain.
In the case of MinGW we can pass gcc as the toolchain name instead of
relying on autodetection.